+++ client/src/main/java/ctbrec/ui/RecordingsTab.java
@@ -6,6 +6,7 @@ import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.net.URL;
+import java.nio.file.NoSuchFileException;
 import java.security.InvalidKeyException;
 import java.security.NoSuchAlgorithmException;
 import java.text.DecimalFormat;
@@ -22,6 +23,8 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ThreadFactory;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.locks.Lock;
+import java.util.concurrent.locks.ReentrantLock;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -52,6 +55,7 @@ import javafx.scene.control.Label;
 import javafx.scene.control.MenuItem;
 import javafx.scene.control.ProgressBar;
 import javafx.scene.control.ScrollPane;
+import javafx.scene.control.SelectionMode;
 import javafx.scene.control.Tab;
 import javafx.scene.control.TableCell;
 import javafx.scene.control.TableColumn;
@@ -91,6 +95,7 @@ public class RecordingsTab extends Tab implements TabSelectionListener {
     ContextMenu popup;
     ProgressBar spaceLeft;
     Label spaceLabel;
+    Lock recordingsLock = new ReentrantLock();
 
     public RecordingsTab(String title, Recorder recorder, Config config, List<Site> sites) {
         super(title);
@@ -114,6 +119,7 @@ public class RecordingsTab extends Tab implements TabSelectionListener {
         BorderPane.setMargin(scrollPane, new Insets(5));
 
         table.setEditable(false);
+        table.getSelectionModel().setSelectionMode(SelectionMode.MULTIPLE);
         TableColumn<JavaFxRecording, String> name = new TableColumn<>("Model");
         name.setPrefWidth(200);
         name.setCellValueFactory(new PropertyValueFactory<JavaFxRecording, String>("modelName"));
@@ -162,14 +168,12 @@ public class RecordingsTab extends Tab implements TabSelectionListener {
                             setStyle(null);
                         } else {
                             setText(StringUtil.formatSize(sizeInByte));
+                            setStyle("-fx-alignment: CENTER-RIGHT;");
                             if(Objects.equals(System.getenv("CTBREC_DEV"), "1")) {
                                 int row = this.getTableRow().getIndex();
                                 JavaFxRecording rec = tableViewProperty().get().getItems().get(row);
                                 if(!rec.valueChanged() && rec.getStatus() == STATUS.RECORDING) {
                                     setStyle("-fx-alignment: CENTER-RIGHT; -fx-background-color: red");
-                                } else {
-                                    setStyle("-fx-alignment: CENTER-RIGHT;");
-                                    //setStyle(null);
                                 }
                             }
                         }
@@ -182,9 +186,9 @@ public class RecordingsTab extends Tab implements TabSelectionListener {
         table.getColumns().addAll(name, date, status, progress, size);
         table.setItems(observableRecordings);
         table.addEventHandler(ContextMenuEvent.CONTEXT_MENU_REQUESTED, event -> {
-            Recording recording = table.getSelectionModel().getSelectedItem();
-            if(recording != null) {
-                popup = createContextMenu(recording);
+            List<JavaFxRecording> recordings = table.getSelectionModel().getSelectedItems();
+            if(recordings != null && !recordings.isEmpty()) {
+                popup = createContextMenu(recordings);
                 if(!popup.getItems().isEmpty()) {
                     popup.show(table, event.getScreenX(), event.getScreenY());
                 }
@@ -205,13 +209,15 @@ public class RecordingsTab extends Tab implements TabSelectionListener {
             }
         });
         table.addEventFilter(KeyEvent.KEY_PRESSED, event -> {
-            JavaFxRecording recording = table.getSelectionModel().getSelectedItem();
-            if (recording != null) {
+            List<JavaFxRecording> recordings = table.getSelectionModel().getSelectedItems();
+            if (recordings != null && !recordings.isEmpty()) {
                 if (event.getCode() == KeyCode.DELETE) {
-                    delete(recording);
+                    if(recordings.size() > 1 || recordings.get(0).getStatus() == STATUS.FINISHED) {
+                        delete(recordings);
+                    }
                 } else if (event.getCode() == KeyCode.ENTER) {
-                    if(recording.getStatus() == STATUS.FINISHED) {
-                        play(recording);
+                    if(recordings.get(0).getStatus() == STATUS.FINISHED) {
+                        play(recordings.get(0));
                     }
                 }
             }
@@ -275,23 +281,28 @@ public class RecordingsTab extends Tab implements TabSelectionListener {
             return;
         }
 
-        for (Iterator<JavaFxRecording> iterator = observableRecordings.iterator(); iterator.hasNext();) {
-            JavaFxRecording old = iterator.next();
-            if (!recordings.contains(old)) {
-                // remove deleted recordings
-                iterator.remove();
+        recordingsLock.lock();
+        try {
+            for (Iterator<JavaFxRecording> iterator = observableRecordings.iterator(); iterator.hasNext();) {
+                JavaFxRecording old = iterator.next();
+                if (!recordings.contains(old)) {
+                    // remove deleted recordings
+                    iterator.remove();
+                }
             }
-        }
-        for (JavaFxRecording recording : recordings) {
-            if (!observableRecordings.contains(recording)) {
-                // add new recordings
-                observableRecordings.add(recording);
-            } else {
-                // update existing ones
-                int index = observableRecordings.indexOf(recording);
-                JavaFxRecording old = observableRecordings.get(index);
-                old.update(recording);
+            for (JavaFxRecording recording : recordings) {
+                if (!observableRecordings.contains(recording)) {
+                    // add new recordings
+                    observableRecordings.add(recording);
+                } else {
+                    // update existing ones
+                    int index = observableRecordings.indexOf(recording);
+                    JavaFxRecording old = observableRecordings.get(index);
+                    old.update(recording);
+                }
             }
+        } finally {
+            recordingsLock.unlock();
         }
         table.sort();
     }
@@ -316,6 +327,10 @@ public class RecordingsTab extends Tab implements TabSelectionListener {
                         try {
                             spaceTotal = recorder.getTotalSpaceBytes();
                             spaceFree = recorder.getFreeSpaceBytes();
+                            Platform.runLater(() -> spaceLeft.setTooltip(new Tooltip()));
+                        } catch (NoSuchFileException e) {
+                            // recordings dir does not exist
+                            Platform.runLater(() -> spaceLeft.setTooltip(new Tooltip("Recordings directory does not exist")));
                         } catch (IOException e) {
                             LOG.error("Couldn't update free space", e);
                         }
@@ -351,7 +366,7 @@ public class RecordingsTab extends Tab implements TabSelectionListener {
         }
     }
 
-    private ContextMenu createContextMenu(Recording recording) {
+    private ContextMenu createContextMenu(List<JavaFxRecording> recordings) {
         ContextMenu contextMenu = new ContextMenu();
         contextMenu.setHideOnEscape(true);
         contextMenu.setAutoHide(true);
@@ -359,9 +374,9 @@ public class RecordingsTab extends Tab implements TabSelectionListener {
 
         MenuItem openInPlayer = new MenuItem("Open in Player");
         openInPlayer.setOnAction((e) -> {
-            play(recording);
+            play(recordings.get(0));
         });
-        if(recording.getStatus() == STATUS.FINISHED || Config.getInstance().getSettings().localRecording) {
+        if(recordings.get(0).getStatus() == STATUS.FINISHED || Config.getInstance().getSettings().localRecording) {
             contextMenu.getItems().add(openInPlayer);
         }
 
@@ -381,16 +396,16 @@ public class RecordingsTab extends Tab implements TabSelectionListener {
 
         MenuItem deleteRecording = new MenuItem("Delete");
         deleteRecording.setOnAction((e) -> {
-            delete(recording);
+            delete(recordings);
         });
-        if(recording.getStatus() == STATUS.FINISHED) {
+        if(recordings.get(0).getStatus() == STATUS.FINISHED || recordings.size() > 1) {
             contextMenu.getItems().add(deleteRecording);
         }
 
         MenuItem openDir = new MenuItem("Open directory");
         openDir.setOnAction((e) -> {
             String recordingsDir = Config.getInstance().getSettings().recordingsDir;
-            String path = recording.getPath();
+            String path = recordings.get(0).getPath();
             File tsFile = new File(recordingsDir, path);
             new Thread(() -> {
                 DesktopIntegration.open(tsFile.getParent());
@@ -403,14 +418,20 @@ public class RecordingsTab extends Tab implements TabSelectionListener {
         MenuItem downloadRecording = new MenuItem("Download");
         downloadRecording.setOnAction((e) -> {
             try {
-                download(recording);
+                download(recordings.get(0));
             } catch (IOException | ParseException | PlaylistException e1) {
                 showErrorDialog("Error while downloading recording", "The recording could not be downloaded", e1);
                 LOG.error("Error while downloading recording", e1);
             }
         });
-        if (!Config.getInstance().getSettings().localRecording && recording.getStatus() == STATUS.FINISHED) {
+        if (!Config.getInstance().getSettings().localRecording && recordings.get(0).getStatus() == STATUS.FINISHED) {
             contextMenu.getItems().add(downloadRecording);
+        }
+
+        if(recordings.size() > 1) {
+            openInPlayer.setDisable(true);
+            openDir.setDisable(true);
+            downloadRecording.setDisable(true);
         }
 
         return contextMenu;
@@ -497,7 +518,7 @@ public class RecordingsTab extends Tab implements TabSelectionListener {
                 @Override
                 public void run() {
                     boolean started = Player.play(recording);
-                    if(started) {
+                    if(started && Config.getInstance().getSettings().showPlayerStarting) {
                         Platform.runLater(() -> Toast.makeText(getTabPane().getScene(), "Starting Player", 2000, 500, 500));
                     }
                 }
@@ -509,7 +530,7 @@ public class RecordingsTab extends Tab implements TabSelectionListener {
                 @Override
                 public void run() {
                     boolean started = Player.play(url);
-                    if(started) {
+                    if(started && Config.getInstance().getSettings().showPlayerStarting) {
                         Platform.runLater(() -> Toast.makeText(getTabPane().getScene(), "Starting Player", 2000, 500, 500));
                     }
                 }
@@ -518,12 +539,16 @@ public class RecordingsTab extends Tab implements TabSelectionListener {
 
     }
 
-    private void delete(Recording r) {
-        if(r.getStatus() != STATUS.FINISHED) {
-            return;
-        }
+    private void delete(List<JavaFxRecording> recordings) {
         table.setCursor(Cursor.WAIT);
-        String msg = "Delete " + r.getModelName() + "/" + r.getStartDate() + " for good?";
+
+        String msg;
+        if(recordings.size() > 1) {
+            msg = "Delete " + recordings.size() + " recordings for good?";
+        } else {
+            Recording r = recordings.get(0);
+            msg = "Delete " + r.getModelName() + "/" + r.getStartDate() + " for good?";
+        }
         AutosizeAlert confirm = new AutosizeAlert(AlertType.CONFIRMATION, msg, YES, NO);
         confirm.setTitle("Delete recording?");
         confirm.setHeaderText(msg);
@@ -533,14 +558,26 @@ public class RecordingsTab extends Tab implements TabSelectionListener {
             Thread deleteThread = new Thread() {
                 @Override
                 public void run() {
+                    recordingsLock.lock();
                     try {
-                        recorder.delete(r);
-                        Platform.runLater(() -> observableRecordings.remove(r));
-                    } catch (IOException | InvalidKeyException | NoSuchAlgorithmException | IllegalStateException e1) {
-                        LOG.error("Error while deleting recording", e1);
-                        showErrorDialog("Error while deleting recording", "Recording not deleted", e1);
+                        List<Recording> deleted = new ArrayList<>();
+                        for (Iterator<JavaFxRecording> iterator = recordings.iterator(); iterator.hasNext();) {
+                            JavaFxRecording r =  iterator.next();
+                            if(r.getStatus() != STATUS.FINISHED) {
+                                continue;
+                            }
+                            try {
+                                recorder.delete(r);
+                                deleted.add(r);
+                            } catch (IOException | InvalidKeyException | NoSuchAlgorithmException | IllegalStateException e1) {
+                                LOG.error("Error while deleting recording", e1);
+                                showErrorDialog("Error while deleting recording", "Recording not deleted", e1);
+                            }
+                        }
+                        observableRecordings.removeAll(deleted);
                     } finally {
-                        table.setCursor(Cursor.DEFAULT);
+                        recordingsLock.unlock();
+                        Platform.runLater(() -> table.setCursor(Cursor.DEFAULT));
                     }
                 }
             };