+++ common/src/main/java/ctbrec/Settings.java
@@ -30,6 +30,7 @@ public class Settings {
     }
 
     public boolean singlePlayer = true;
+    public boolean showPlayerStarting = false;
     public boolean localRecording = true;
     public int httpPort = 8080;
     public int httpTimeout = 10000;
@@ -42,10 +43,12 @@ public class Settings {
     public String postProcessing = "";
     public String username = ""; // chaturbate username TODO maybe rename this onetime
     public String password = ""; // chaturbate password TODO maybe rename this onetime
+    public String chaturbateBaseUrl = "https://chaturbate.com";
     public String bongaUsername = "";
     public String bongaPassword = "";
     public String mfcUsername = "";
     public String mfcPassword = "";
+    public String mfcBaseUrl = "https://www.myfreecams.com";
     public String camsodaUsername = "";
     public String camsodaPassword = "";
     public String cam4Username;