+++ common/src/main/java/ctbrec/sites/chaturbate/Chaturbate.java
@@ -10,6 +10,8 @@ import java.util.List;
 import java.util.Objects;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -43,14 +45,14 @@ import okhttp3.Response;
 public class Chaturbate extends AbstractSite {
 
     private static final transient Logger LOG = LoggerFactory.getLogger(Chaturbate.class);
-    public static final String BASE_URI = "https://chaturbate.com";
-    public static final String AFFILIATE_LINK = BASE_URI + "/in/?track=default&tour=grq0&campaign=55vTi";
-    public static final String REGISTRATION_LINK = BASE_URI + "/in/?track=default&tour=g4pe&campaign=55vTi";
+    static String baseUrl = "https://chaturbate.com";
+    public static final String AFFILIATE_LINK = "https://chaturbate.com/in/?track=default&tour=grq0&campaign=55vTi";
+    public static final String REGISTRATION_LINK = "https://chaturbate.com/in/?track=default&tour=g4pe&campaign=55vTi";
     private ChaturbateHttpClient httpClient;
 
     @Override
     public void init() throws IOException {
-
+        baseUrl = Config.getInstance().getSettings().chaturbateBaseUrl;
     }
 
     @Override
@@ -60,7 +62,7 @@ public class Chaturbate extends AbstractSite {
 
     @Override
     public String getBaseUrl() {
-        return "https://chaturbate.com";
+        return baseUrl;
     }
 
     @Override
@@ -137,7 +139,7 @@ public class Chaturbate extends AbstractSite {
 
     @Override
     public List<Model> search(String q) throws IOException, InterruptedException {
-        String url = BASE_URI + "?keywords=" + URLEncoder.encode(q, "utf-8");
+        String url = baseUrl + "?keywords=" + URLEncoder.encode(q, "utf-8");
         List<Model> result = new ArrayList<>();
 
         // search online models
@@ -153,7 +155,7 @@ public class Chaturbate extends AbstractSite {
 
         // since chaturbate does not return offline models, we at least try, if the profile page
         // exists for the search string
-        url = BASE_URI + '/' + q;
+        url = baseUrl + '/' + q;
         req = new Request.Builder()
                 .url(url)
                 .addHeader("User-Agent", Config.getInstance().getSettings().httpUserAgent)
@@ -238,7 +240,7 @@ public class Chaturbate extends AbstractSite {
                 .add("bandwidth", "high")
                 .build();
         Request req = new Request.Builder()
-                .url("https://chaturbate.com/get_edge_hls_url_ajax/")
+                .url(getBaseUrl() + "/get_edge_hls_url_ajax/")
                 .post(body)
                 .addHeader("X-Requested-With", "XMLHttpRequest")
                 .build();
@@ -339,5 +341,16 @@ public class Chaturbate extends AbstractSite {
     public boolean credentialsAvailable() {
         String username = Config.getInstance().getSettings().username;
         return username != null && !username.trim().isEmpty();
+    }
+
+    @Override
+    public Model createModelFromUrl(String url) {
+        Matcher m = Pattern.compile("https?://.*?chaturbate.com(?:/p)?/([^/]*?)/?").matcher(url);
+        if(m.matches()) {
+            String modelName = m.group(1);
+            return createModel(modelName);
+        } else {
+            return super.createModelFromUrl(url);
+        }
     }
 }