+++ common/src/main/java/ctbrec/sites/mfc/MyFreeCamsClient.java
@@ -85,7 +85,7 @@ public class MyFreeCamsClient {
 
     public void start() throws IOException {
         running = true;
-        serverConfig = new ServerConfig(mfc.getHttpClient());
+        serverConfig = new ServerConfig(mfc);
         List<String> websocketServers = new ArrayList<String>(serverConfig.wsServers.keySet());
         String server = websocketServers.get((int) (Math.random()*websocketServers.size()));
         String wsUrl = "ws://" + server + ".myfreecams.com:8080/fcsl";