+++ common/src/main/java/ctbrec/sites/mfc/MyFreeCamsModel.java
@@ -99,7 +99,7 @@ public class MyFreeCamsModel extends AbstractModel {
     }
 
     private MasterPlaylist getMasterPlaylist() throws IOException, ParseException, PlaylistException {
-        if(hlsUrl == null) {
+        if(getHlsUrl() == null) {
             throw new IllegalStateException("Stream url unknown");
         }
         LOG.trace("Loading master playlist {}", hlsUrl);
@@ -117,6 +117,14 @@ public class MyFreeCamsModel extends AbstractModel {
         }
     }
 
+    private String getHlsUrl() {
+        if(hlsUrl == null) {
+            MyFreeCams mfc = (MyFreeCams) getSite();
+            mfc.getClient().update(this);
+        }
+        return hlsUrl;
+    }
+
     @Override
     public void invalidateCacheEntries() {
         resolution = null;
@@ -124,7 +132,7 @@ public class MyFreeCamsModel extends AbstractModel {
 
     @Override
     public void receiveTip(int tokens) throws IOException {
-        String tipUrl = MyFreeCams.BASE_URI + "/php/tip.php";
+        String tipUrl = MyFreeCams.baseUrl + "/php/tip.php";
         String initUrl = tipUrl + "?request=tip&username="+getName()+"&broadcaster_id="+getUid();
         Request req = new Request.Builder().url(initUrl).build();
         try(Response resp = site.getHttpClient().execute(req)) {